refactor(billing): splits monolithic billing router into focused modules
