--- core/ai/src/observability.ts
@@ -1,8 +1,0 @@
-import type { PostHog } from "@core/posthog/server";
-
-export type AiObservabilityContext = {
-   posthog: PostHog;
-   distinctId: string;
-   promptName?: string;
-   promptVersion?: number;
-};
