--- modules/billing/src/contracts/billing.ts
@@ -1,1 +1,0 @@
-export {};
