+++ modules/billing/src/router/customer-portal.ts
@@ -1,9 +1,22 @@
 import { fromPromise } from "neverthrow";
+import { billingContract } from "@montte/hyprpay/contract";
+import { implementerInternal } from "@orpc/server";
 import { getDomain } from "@core/environment/helpers";
 import { WebAppError } from "@core/logging/errors";
-import { billingImpl } from "./_implementer";
+import { protectedProcedure } from "@core/orpc/server";
+import type {
+   ORPCContext,
+   ORPCContextWithOrganization,
+} from "@core/orpc/server";
 
-export const createSession = billingImpl.customerPortal.createSession.handler(
+const def = protectedProcedure["~orpc"];
+const impl = implementerInternal<
+   typeof billingContract.customerPortal,
+   ORPCContext,
+   ORPCContextWithOrganization
+>(billingContract.customerPortal, def.config, [...def.middlewares]);
+
+export const createSession = impl.createSession.handler(
    async ({ context, input }) => {
       const contactResult = await fromPromise(
          context.db.query.contacts.findFirst({
