+++ modules/classification/src/ai/classify-batch.ts
@@ -5,8 +5,10 @@ import { AppError } from "@core/logging/errors";
 import { promptsClient } from "@core/posthog/server";
 import { CLASSIFICATION_PROMPTS } from "../constants";
 import { flashModel } from "@core/ai/models";
-import { createPosthogAiMiddleware } from "@core/ai/middleware";
-import type { AiObservabilityContext } from "@core/ai/observability";
+import {
+   type AiObservabilityContext,
+   createPosthogAiMiddleware,
+} from "@core/ai/middleware";
 
 const MAX_BATCH_SIZE = 20;
